Upgrade to Webpack5 + TS config
