+++ package.json
@@ -27,28 +27,33 @@
     "@types/anchor-js": "^4.2.0",
     "@types/debounce": "^1.2.0",
     "@types/jest": "^26.0.19",
-    "@types/node": "^12.19.9",
+    "@types/node": "^12.20.19",
+    "@types/semver": "^7.3.8",
+    "@types/webpack": "^5.28.0",
+    "@types/webpack-dev-server": "^3.11.5",
     "@typescript-eslint/eslint-plugin": "^4.11.0",
     "@typescript-eslint/parser": "^4.11.0",
     "babel-jest": "^26.6.3",
     "babel-loader": "^8.2.2",
-    "eslint": "^7.16.0",
-    "eslint-config-preact": "^1.1.3",
-    "eslint-config-prettier": "^6.15.0",
-    "eslint-plugin-import": "^2.22.1",
+    "eslint": "^7.32.0",
+    "eslint-config-preact": "^1.1.4",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-import": "^2.23.4",
     "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-prettier": "^3.3.0",
-    "execa": "^4.1.0",
-    "fs-extra": "^9.0.1",
+    "eslint-plugin-prettier": "^3.4.0",
+    "execa": "^5.1.1",
+    "fs-extra": "^10.0.0",
     "jest": "^26.6.3",
     "listr": "^0.14.3",
     "prettier": "2.0.2",
     "semver": "^7.3.4",
     "ts-jest": "^26.4.4",
-    "ts-loader": "^6.2.2",
+    "ts-loader": "^9.2.5",
+    "ts-node": "^10.1.0",
     "typescript": "^3.9.7",
-    "webpack": "^4.44.2",
+    "webpack": "^5.49.0",
     "webpack-cli": "^3.3.12",
+    "webpack-dev-server": "^3.11.2",
     "yargs": "^15.4.1"
   }
 }
